pyproject.toml
[tool.poetry]
name = "llms-for-ts"
version = "0.1.0"
description = "Undergraduate thesis repository to evaluate the usage of LLMs as zero-shot time series forecasting models."
authors = ["VittorioRossi <rossi.vittori@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
kaggle = "^1.6.12"
python-dotenv = "^1.0.1"
click = "^8.1.7"
pandas = "^2.2.2"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
